created a folders for the competition, and moved the IsaacCompetition file into this folder
